requirements.txt
pandas==2.0.3 # requirement of opensearch-py-ml
deprecated~=1.2.14  # requirement of opensearch-py-ml
accelerate~=0.26.1  # requirement of opensearch-py-ml
ollama

Bio==1.6.2
datasets==2.16.1
fastapi==0.109.0
langchain==0.1.7
opensearch_py_ml==1.1.0
pydantic==2.5.3
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.30.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
uvicorn==0.25.0

sentence_transformers # for ingestor.py